Updated methods to work with what the frontend is sending.
Changed GET CheckExistingStudyModule to return a key value pair of a boolean value and either a topicId or null. Updated GenerateStudyModuleFromTopic method to include the mindMapiId when it generates a study tool so that it saves a valid TopicId in the StudyTools table and that fixed the button state on the frontend.